🧪 Move `hashivault` tests to a separate module
